requirements.txt
quart
aiohttp
hypercorn
python-dotenv
pydantic
g4f==0.5.1.2
colorama